clean up
https://openedx.atlassian.net/browse/MA-2557
